add rtl to Box RTL story args
